added helper methods
