+++ Plugin/Movie/ScriptEngine/ScriptEngineMovieRunnerExterns.cs
@@ -1,0 +1,162 @@
+using System.Collections.Generic;
+using System.Linq;
+using UniTASPlugin.Movie.ScriptEngine.EngineMethods.Exceptions;
+using UniTASPlugin.Movie.ScriptEngine.LowLevelEngine;
+using UniTASPlugin.Movie.ScriptEngine.MovieModels.Script;
+using UniTASPlugin.Movie.ScriptEngine.OpCodes;
+using UniTASPlugin.Movie.ScriptEngine.OpCodes.Method;
+using UniTASPlugin.Movie.ScriptEngine.OpCodes.RegisterSet;
+using UniTASPlugin.Movie.ScriptEngine.ValueTypes;
+
+namespace UniTASPlugin.Movie.ScriptEngine;
+
+public partial class ScriptEngineMovieRunner
+{
+    public int RegisterConcurrentMethod(string methodName, bool preUpdate,
+        IEnumerable<IEnumerable<ValueType>> defaultArgs)
+    {
+        if (methodName == null) return -1;
+
+        var foundDefinedMethod = _mainScript.Methods.ToList().Find(x => x.Name == methodName);
+        var externFound = foundDefinedMethod == null ? _externalMethods.ToList().Find(x => x.Name == methodName) : null;
+
+        if (foundDefinedMethod == null && externFound == null)
+        {
+            return -1;
+        }
+
+        // check if arg count match
+        var argCount = foundDefinedMethod == null
+            ? externFound.ArgCount
+            : foundDefinedMethod.OpCodes.Count(x => x is PopArgOpCode);
+
+        var argTuples = defaultArgs as IEnumerable<ValueType>[] ?? defaultArgs.ToArray();
+        if (argCount != argTuples.Count())
+        {
+            throw new MethodArgCountNotMatching(argCount.ToString(), argTuples.Count().ToString());
+        }
+
+        ScriptModel runnerScript;
+        if (foundDefinedMethod != null)
+        {
+            var mainMethod = new List<OpCodeBase>();
+            foreach (var argTuple in argTuples)
+            {
+                foreach (var arg in argTuple)
+                {
+                    mainMethod.Add(new ConstToRegisterOpCode(RegisterType.Temp0, arg));
+                    mainMethod.Add(new PushArgOpCode(RegisterType.Temp0));
+                }
+            }
+
+            mainMethod.Add(new GotoMethodOpCode(foundDefinedMethod.Name));
+
+            runnerScript = new ScriptModel(new(null, mainMethod),
+                new[] { foundDefinedMethod });
+        }
+        else
+        {
+            var wrapperMethod = new List<OpCodeBase>();
+            foreach (var argTuple in argTuples)
+            {
+                foreach (var arg in argTuple)
+                {
+                    wrapperMethod.Add(new ConstToRegisterOpCode(RegisterType.Temp0, arg));
+                    wrapperMethod.Add(new PushArgOpCode(RegisterType.Temp0));
+                }
+            }
+
+            wrapperMethod.Add(new GotoMethodOpCode(externFound.Name));
+
+            wrapperMethod.Add(new GotoMethodOpCode(externFound.Name));
+            runnerScript = new ScriptModel(new ScriptMethodModel(null, wrapperMethod), new ScriptMethodModel[0]);
+        }
+
+        var engine = new ScriptEngineLowLevelEngine(runnerScript, _externalMethods);
+
+        if (preUpdate)
+        {
+            _concurrentRunnersPreUpdate.Add(engine);
+            // actually run it since otherwise it will skip a frame
+            engine.ExecUntilStop(this);
+            if (engine.FinishedExecuting)
+            {
+                engine.Reset();
+            }
+        }
+        else
+        {
+            _concurrentRunnersPostUpdate.Add(engine);
+        }
+
+        return preUpdate
+            ? _concurrentRunnersPreUpdate.Last().GetHashCode()
+            : _concurrentRunnersPostUpdate.Last().GetHashCode();
+    }
+
+    public void UnregisterConcurrentMethod(int hashCode, bool preUpdate)
+    {
+        if (preUpdate)
+        {
+            var foundRunner = _concurrentRunnersPreUpdate.FindIndex(x => x.GetHashCode() == hashCode);
+            if (foundRunner < 0) return;
+            _concurrentRunnersPreUpdate.RemoveAt(foundRunner);
+        }
+        else
+        {
+            var foundRunner = _concurrentRunnersPostUpdate.FindIndex(x => x.GetHashCode() == hashCode);
+            if (foundRunner < 0) return;
+            _concurrentRunnersPostUpdate.RemoveAt(foundRunner);
+        }
+    }
+
+    private void ConcurrentRunnersPreUpdate()
+    {
+        foreach (var runner in _concurrentRunnersPreUpdate)
+        {
+            runner.ExecUntilStop(this);
+            if (runner.FinishedExecuting)
+            {
+                runner.Reset();
+            }
+        }
+    }
+
+    private void ConcurrentRunnersPostUpdate()
+    {
+        foreach (var runner in _concurrentRunnersPostUpdate)
+        {
+            runner.ExecUntilStop(this);
+            if (runner.FinishedExecuting)
+            {
+                runner.Reset();
+            }
+        }
+    }
+
+    public void UpdateHeldKey(bool press, int keyCode)
+    {
+        var env = _virtualEnvironmentService.GetVirtualEnv();
+        if (press)
+        {
+            env.InputState.KeyboardState.Keys.Add(keyCode);
+        }
+        else
+        {
+            env.InputState.KeyboardState.Keys.RemoveAt(keyCode);
+        }
+    }
+
+    public void ClearHeldKeys()
+    {
+        var env = _virtualEnvironmentService.GetVirtualEnv();
+        env.InputState.KeyboardState.Keys.Clear();
+    }
+
+    public void MoveMouse(int x, int y)
+    {
+        var env = _virtualEnvironmentService.GetVirtualEnv();
+        env.InputState.MouseState.XPos = x;
+        env.InputState.MouseState.YPos = y;
+    }
+}
