+++ Plugin/NInjectModules/GameEnvironmentModule.cs
@@ -1,4 +1,5 @@
-﻿using Ninject.Modules;
+﻿using Ninject.Extensions.Factory;
+using Ninject.Modules;
 using UniTASPlugin.GameEnvironment;
 
 namespace UniTASPlugin.NInjectModules;
@@ -8,5 +9,6 @@ public class GameEnvironmentModule : NinjectModule
     public override void Load()
     {
         Bind<VirtualEnvironment>().ToSelf().InSingletonScope();
+        Bind<IVirtualEnvironmentService>().ToFactory();
     }
 }
